@@ -18,18 +18,20 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
-  http: 0.13.0
+  http: ^0.13.0
   http_parser: ^4.0.0
   provider: ^5.0.0
-  firebase_storage_web: ^1.0.0
-  cloud_firestore: ^1.0.0 
-  firebase_core: ^1.0.0
+  firebase_storage_web: ^1.0.1
+  cloud_firestore: ^1.0.1 
+  firebase_core: ^1.0.1
   firebase_storage: ^8.0.0
+  image_picker: ^0.7.2+1
+  permission_handler: ^6.1.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
