@@ -1,5 +1,5 @@
 class NoteList {
-  final String id, title, subtitle;
+  final String? id, title, subtitle;
   NoteList({this.id, this.title, this.subtitle});
 
   NoteList.fromMap(Map<String, dynamic> data, String id)
